Refactoring. Hooray! AutoTrackerCore is eventually separated
