+++ src/org/opensourcephysics/cabrillo/tracker/AutoTrackerFeedback.java
@@ -1,6 +1,9 @@
 package org.opensourcephysics.cabrillo.tracker;
 
-import org.opensourcephysics.cabrillo.tracker.AutoTracker.KeyFrame;
+import org.opensourcephysics.cabrillo.tracker.AutoTrackerCore.KeyFrame;
+import org.opensourcephysics.cabrillo.tracker.AutoTrackerCore.FrameData;
+
+import java.awt.image.BufferedImage;
 
 /**
  * @author Nikolai Avdeev aka NickKolok
@@ -10,4 +13,12 @@ public class AutoTrackerFeedback {
 
 	public void onBeforeAddKeyframe(double x, double y){}
 	public void onAfterAddKeyframe(KeyFrame keyFrame){}
+
+	public void onSetTrack(){}
+
+	public void onTrackUnbind(TTrack track){}
+	public void onTrackBind(TTrack track){}
+
+	public void onTemplateSetForFrame(AutoTrackerCore.FrameData frameData, BufferedImage template){}
+
 }
